fixed mac asset renaming logic

+++ src/misc.ts
@@ -32,25 +32,39 @@ export function findCurrentAssets(platform: string, arch: string, tauri: TauriPr
             if (arch === "intel") {
                 let macPath = project_path + "/src-tauri/target/x86_64-apple-darwin/release/bundle/macos/";
 
-                fs.renameSync(macPath + `${tauri.package.productName}.app`, macPath + `${tauri.package.productName}_x86_64.app`);
-                fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz`, macPath + `${tauri.package.productName}_x86_64.app.tar.gz`);
-                fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`, macPath + `${tauri.package.productName}_x86_64.app.tar.gz.sig`);
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app`, macPath + `${tauri.package.productName}_x86_64.app`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app`);
+                }
 
-                assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app`);
-                assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app.tar.gz`);
-                assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app.tar.gz.sig`);
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app.tar.gz`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz`, macPath + `${tauri.package.productName}_x86_64.app.tar.gz`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app.tar.gz`);
+                }
+
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`, macPath + `${tauri.package.productName}_x86_64.app.tar.gz.sig`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_x86_64.app.tar.gz.sig`);
+                }
 
                 altArch = "darwin-x86_64";
             } else if (arch === "silicon") {
                 let macPath = project_path + "/src-tauri/target/aarch64-apple-darwin/release/bundle/macos/";
 
-                fs.renameSync(macPath + `${tauri.package.productName}.app`, macPath + `${tauri.package.productName}_aarch64.app`);
-                fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz`, macPath + `${tauri.package.productName}_aarch64.app.tar.gz`);
-                fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`, macPath + `${tauri.package.productName}_aarch64.app.tar.gz.sig`);
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app`, macPath + `${tauri.package.productName}_aarch64.app`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app`);
+                }
 
-                assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app`);
-                assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app.tar.gz`);
-                assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app.tar.gz.sig`);
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app.tar.gz`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz`, macPath + `${tauri.package.productName}_aarch64.app.tar.gz`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app.tar.gz`);
+                }
+
+                if (fs.existsSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`)) {
+                    fs.renameSync(macPath + `${tauri.package.productName}.app.tar.gz.sig`, macPath + `${tauri.package.productName}_aarch64.app.tar.gz.sig`);
+                    assetPaths.push(macPath + `${tauri.package.productName}_aarch64.app.tar.gz.sig`);
+                }
 
                 altArch = "darwin-aarch64";
             }
